+++ 30_days_exercise/src/Controller/MainMenu/Start.java
@@ -1,0 +1,8 @@
+package Controller.MainMenu;
+
+public class Start {
+    public static void main(String[] args) {
+    Main toMain = new Main();
+    toMain.openingMenu();
+    }
+}
